Reworked exoand and heredoc handles
