+++ out
@@ -1,0 +1,2 @@
+hi
+hello
