+++ sources/main.c
@@ -6,7 +6,7 @@
 /*   By: nzharkev <nzharkev@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/26 15:40:55 by nzharkev          #+#    #+#             */
-/*   Updated: 2024/12/23 15:46:04 by nzharkev         ###   ########.fr       */
+/*   Updated: 2024/12/23 18:18:06 by nzharkev         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
