+++ sources/parser/expand.c
@@ -6,16 +6,16 @@
 /*   By: nzharkev <nzharkev@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/14 10:58:12 by nzharkev          #+#    #+#             */
-/*   Updated: 2024/12/23 14:59:17 by nzharkev         ###   ########.fr       */
+/*   Updated: 2024/12/23 18:22:22 by nzharkev         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minishell.h"
 
-int	handle_expand(t_shell *mini, t_cmd **cmd);
+int			handle_expand(t_shell *mini, t_cmd **cmd);
 static int	this_is_space(t_cmd **cmd, t_expand *arg, char **expan);
-static int	expand_while(t_shell *mini, t_cmd **cmd, t_expand *arg, char **expan);
-static int	seg_no_quotes(t_shell *mini, t_cmd **cmd, t_expand *arg, char **expan);
+static int	exp_while(t_shell *mini, t_cmd **cmd, t_expand *arg, char **expan);
+static int	s_unquoted(t_shell *mini, t_cmd **cmd, t_expand *arg, char **expan);
 static int	hd_quoted(t_shell *mini, t_cmd **cmd, t_expand *arg, char **expan);
 
 int	handle_expand(t_shell *mini, t_cmd **cmd)
@@ -26,14 +26,14 @@ int	handle_expand(t_shell *mini, t_cmd **cmd)
 	expan = NULL;
 	if (init_expansion(&arg, &expan))
 		return (1);
-	if (expand_while(mini, cmd, &arg, &expan))
+	if (exp_while(mini, cmd, &arg, &expan))
 		return (1);
 	free((*cmd)->seg);
 	(*cmd)->seg = expan;
 	return (0);
 }
 
-static int	expand_while(t_shell *mini, t_cmd **cmd, t_expand *arg, char **expan)
+static int	exp_while(t_shell *mini, t_cmd **cmd, t_expand *arg, char **expan)
 {
 	while ((*cmd)->seg[arg->i])
 	{
@@ -52,7 +52,7 @@ static int	expand_while(t_shell *mini, t_cmd **cmd, t_expand *arg, char **expan)
 		}
 		else
 		{
-			arg->i = seg_no_quotes(mini, cmd, arg, expan);
+			arg->i = s_unquoted(mini, cmd, arg, expan);
 			if (arg->i == -1)
 				return (1);
 		}
@@ -64,11 +64,7 @@ static int	this_is_space(t_cmd **cmd, t_expand *arg, char **expan)
 {
 	char	*temp;
 
-	temp = ft_strdup(*expan);
-	if (!temp)
-		return (-1);
-	free(*expan);
-	// temp = *expan;
+	temp = *expan;
 	*expan = ft_strjoin_char(temp, (*cmd)->seg[arg->i]);
 	if (!*expan)
 		return (-1);
@@ -83,38 +79,29 @@ static int	hd_quoted(t_shell *mini, t_cmd **cmd, t_expand *arg, char **expan)
 
 	if ((*cmd)->seg[arg->i] == '<' && (*cmd)->seg[arg->i + 1] == '<')
 	{
-		temp = ft_strdup(*expan);
-		free(*expan);
-		// temp = *expan;
-		arg->i = we_have_heredoc(arg, (*cmd)->seg);
+		temp = *expan;
+		arg->i = we_have_heredoc(arg, (*cmd)->seg, 0);
 		*expan = ft_strjoin(temp, arg->value);
 		free(temp);
 	}
 	if ((*cmd)->seg[arg->i] == '\'' || (*cmd)->seg[arg->i] == '"')
 	{
-		temp = ft_strdup(*expan);
-		free(*expan);
-		// temp = *expan;
+		temp = *expan;
 		arg->i = in_quotes(mini, (*cmd)->seg, arg->i, arg);
 		*expan = ft_strjoin(temp, arg->value);
 		free(temp);
 	}
-	// free(arg->value);
 	return (arg->i);
 }
 
-static int	seg_no_quotes(t_shell *mini, t_cmd **cmd, t_expand *arg, char **expan)
+static int	s_unquoted(t_shell *mini, t_cmd **cmd, t_expand *arg, char **expan)
 {
 	char	*temp;
 
 	temp = ft_strdup(*expan);
 	free(*expan);
-	// temp = *expan;
 	arg->i = no_quotes(mini, *cmd, arg->i, arg);
-	// printf("temp: {%s}\nvalue: {%s}\n", temp, arg->value);
 	*expan = ft_strjoin(temp, arg->value);
-	// printf("exapan: {%s}\n", *expan);
 	free(temp);
-	free(arg->value);
 	return (arg->i);
 }
