+++ sources/parser/expand_utils.c
@@ -5,86 +5,105 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: nzharkev <nzharkev@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/12/02 15:12:11 by nzharkev          #+#    #+#             */
-/*   Updated: 2024/12/23 11:30:36 by nzharkev         ###   ########.fr       */
+/*   Created: 2024/12/20 14:44:51 by nzharkev          #+#    #+#             */
+/*   Updated: 2024/12/23 17:37:00 by nzharkev         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minishell.h"
 
-int			in_quotes(t_shell *mini, char *str, int i, t_expand *arg);
-static int	handle_dollar(t_shell *mini, t_expand *arg, char *str);
-static int	handle_quotes(t_expand *arg, char *str);
-int			we_have_heredoc(t_expand *arg, char *str);
+int			we_have_dollar(t_shell *mini, t_expand *arg, char *str);
+int			oh_a_dollar(t_shell *mini, char *str, char **expan, t_expand *arg);
+static int	finalize_expand(t_shell *mini, t_vdata *data, t_expand *arg);
+static int	handle_new_expand(char *temp, char **expan);
+static void	init_vdata(t_vdata *data, char **expan, char *temp, char *name);
 
-int	in_quotes(t_shell *mini, char *str, int i, t_expand *arg)
+int	we_have_dollar(t_shell *mini, t_expand *arg, char *str)
 {
+	int		s_exp;
 	char	*temp;
+	char	*new_res;
 
-	if (the_arg(arg, i))
+	s_exp = arg->i;
+	temp = ft_strdup("");
+	if (!temp)
 		return (-1);
-	temp = arg->value;
-	arg->value = ft_strjoin_char(temp, str[arg->i]);
+	arg->start = arg->i;
+	arg->i = oh_a_dollar(mini, str, &temp, arg);
+	if (arg->i == -1)
+	{
+		free(temp);
+		return (-1);
+	}
+	new_res = ft_strjoin(arg->value, temp);
+	if (!new_res)
+	{
+		free(temp);
+		return (-1);
+	}
+	free(arg->value);
 	free(temp);
-	what_quote(str, arg);
-	while (str[arg->i])
-	{
-		if (!arg->sgl && !arg->dbl)
-			break ;
-		if (str[arg->i] == '$')
-			arg->i = handle_dollar(mini, arg, str);
-		if ((!arg->sgl && !arg->dbl && (str[arg->i] == '\''
-			|| str[arg->i] == '"')) || ((arg->sgl && str[arg->i] == '\'')
-		|| (arg->dbl && str[arg->i] == '"')))
-			arg->i = handle_quotes(arg, str);
-		else if (add_char(str, arg))
-			return (free(arg->value), -1);
-	}
-	arg->len = ft_strlen(arg->value);
-	return (arg->i);
+	arg->value = new_res;
+	return (0);
 }
 
-int	we_have_heredoc(t_expand *arg, char *str)
-{
-	arg->value = ft_strdup("");
-	while (str[arg->i] == '<' || str[arg->i] == ' ')
-		add_char(str, arg);
-	while (str[arg->i] && (!ft_isspace(str[arg->i])
-			|| !(str[arg->i] == '<') || !(str[arg->i] == '>')))
-		add_char(str, arg);
-	return (arg->i);
-}
-
-static int	handle_dollar(t_shell *mini, t_expand *arg, char *str)
-{
-	if (str[arg->i] == '$' && str[arg->i + 1]
-		&& (str[arg->i + 1] == '\'' || str[arg->i + 1] == '"')
-		&& (!arg->sgl && !arg->dbl))
-			arg->i++;
-	if ((str[arg->i] == '$' && !arg->sgl) && (str[arg->i + 1]
-			&& !ft_isspace(str[arg->i + 1]) && (ft_isalnum(str[arg->i + 1])
-				|| str[arg->i + 1] == '_' || str[arg->i + 1] == '?')))
-	{
-		if (we_have_dollar(mini, arg, str) == -1)
-			return (free(arg->value), -1);
-	}
-	return (arg->i);
-}
-
-static int	handle_quotes(t_expand *arg, char *str)
+int	oh_a_dollar(t_shell *mini, char *str, char **expan, t_expand *arg)
 {
 	char	*temp;
+	char	name[100];
+	int		indx;
+	t_vdata	data;
 
-	if (!arg->sgl && !arg->dbl && (str[arg->i] == '\''
-			|| str[arg->i] == '"'))
-		what_quote(str, arg);
-	else if ((arg->sgl && str[arg->i] == '\'')
-		|| (arg->dbl && str[arg->i] == '"'))
+	temp = ft_strndup(&str[arg->start], arg->i - arg->start);
+	if (!temp)
+		return (-1);
+	if (handle_new_expand(temp, expan))
+		return (-1);
+	arg->i++;
+	indx = 0;
+	while (str[arg->i] && (ft_isalnum(str[arg->i])
+			|| str[arg->i] == '_' || str[arg->i] == '?'))
 	{
-		temp = arg->value;
-		arg->value = ft_strjoin_char(temp, str[arg->i]);
-		what_quote(str, arg);
-		free(temp);
+		if (indx < (int) sizeof(name) - 1)
+			name[indx++] = str[arg->i++];
 	}
-	return (arg->i);
+	name[indx] = '\0';
+	init_vdata(&data, expan, temp, name);
+	if (finalize_expand(mini, &data, arg))
+		return (-1);
+	return (arg->start);
+}
+
+static int	finalize_expand(t_shell *mini, t_vdata *data, t_expand *arg)
+{
+	if (handle_value(mini, data))
+		return (-1);
+	arg->name = ft_strdup(data->name);
+	arg->start += ft_strlen(arg->name) + 1;
+	free(arg->name);
+	return (0);
+}
+
+static int	handle_new_expand(char *temp, char **expan)
+{
+	char	*new_expanded;
+
+	new_expanded = ft_strjoin(*expan, temp);
+	if (!new_expanded)
+	{
+		ft_putendl_fd("malloc failed", 2);
+		return (-1);
+	}
+	free(temp);
+	free(*expan);
+	*expan = new_expanded;
+	return (0);
+}
+
+static void	init_vdata(t_vdata *data, char **expan, char *temp, char *name)
+{
+	data->value = NULL;
+	data->expan = expan;
+	data->temp = temp;
+	data->name = name;
 }
