+++ sources/parser/expand_utils3.c
@@ -1,40 +1,93 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   expand_utils2.c                                    :+:      :+:    :+:   */
+/*   expand_utils3.c                                    :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: nzharkev <nzharkev@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/12/20 14:44:51 by nzharkev          #+#    #+#             */
-/*   Updated: 2024/12/21 20:17:00 by nzharkev         ###   ########.fr       */
+/*   Created: 2024/12/21 18:55:47 by nzharkev          #+#    #+#             */
+/*   Updated: 2024/12/23 17:30:55 by nzharkev         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../../includes/minishell.h"
 
-int	we_have_dollar(t_shell *mini, t_expand *arg, char *str);
-int	oh_its_a_dollar(t_shell *mini, char *str, char **expan, t_expand *arg);
-int	tildes_home(t_shell *mini, char *str, char **expan, t_expand *arg);
-static int	handle_new_expand(char *temp, char **expan);
-static void	init_vdata(t_vdata *data, char **expan, char *temp, char *name);
+int		init_expansion(t_expand *arg, char **expan);
+int		the_arg(t_expand *arg, int i);
+void	what_quote(char *str, t_expand *arg);
+int		handle_question(t_shell *mini, char *str, char **expan, t_expand *arg);
+int		new_result(t_expand *arg, char *temp);
 
-int	we_have_dollar(t_shell *mini, t_expand *arg, char *str)
+int	init_expansion(t_expand *arg, char **expan)
 {
-	int		s_exp;
-	char	*temp;
+	*expan = ft_strdup("");
+	if (!(*expan))
+		return (1);
+	the_arg(arg, 0);
+	if (!arg->value)
+		return (1);
+	free(arg->value);
+	return (0);
+}
+
+int	the_arg(t_expand *arg, int i)
+{
+	arg->sgl = 0;
+	arg->dbl = 0;
+	arg->i = i;
+	arg->start = i;
+	arg->name = NULL;
+	arg->value = ft_strdup("");
+	if (!arg->value)
+	{
+		ft_putendl_fd("Malloc fail", 2);
+		return (1);
+	}
+	return (0);
+}
+
+void	what_quote(char *str, t_expand *arg)
+{
+	if ((arg->sgl == 1 && str[arg->i] == '\'') || (arg->dbl == 1
+			&& str[arg->i] == '"'))
+	{
+		if (str[arg->i] == '\'')
+			arg->sgl = !arg->sgl;
+		else if (str[arg->i] == '\"')
+			arg->dbl = !arg->dbl;
+		arg->i++;
+		return ;
+	}
+	if (str[arg->i] == '\'' && arg->sgl == 0)
+	{
+		arg->sgl = 1;
+		arg->i++;
+		return ;
+	}
+	if (str[arg->i] == '"' && arg->dbl == 0)
+	{
+		arg->dbl = 1;
+		arg->i++;
+		return ;
+	}
+}
+
+int	handle_question(t_shell *mini, char *str, char **expan, t_expand *arg)
+{
+	int	cont;
+
+	arg->start = arg->i;
+	cont = arg->start;
+	if (str[arg->i] == '$')
+		arg->i = oh_a_dollar(mini, str, expan, arg);
+	arg->i = cont + 2;
+	return (arg->i);
+}
+
+int	new_result(t_expand *arg, char *temp)
+{
 	char	*new_res;
 
-	s_exp = arg->i;
-	temp = ft_strdup("");
-	if (!temp)
-		return (-1);
-	arg->start = arg->i;
-	arg->i = oh_its_a_dollar(mini, str, &temp, arg);
-	if (arg->i == -1)
-	{
-		free(temp);
-		return (-1);
-	}
 	new_res = ft_strjoin(arg->value, temp);
 	if (!new_res)
 	{
@@ -45,81 +98,4 @@ int	we_have_dollar(t_shell *mini, t_expand *arg, char *str)
 	free(temp);
 	arg->value = new_res;
 	return (0);
-}
-
-int	oh_its_a_dollar(t_shell *mini, char *str, char **expan, t_expand *arg)
-{
-	char	*temp;
-	char	name[100];
-	int		indx;
-	t_vdata	data;
-
-	temp = ft_strndup(&str[arg->start], arg->i - arg->start);
-	if (!temp)
-		return (-1);
-	if (handle_new_expand(temp, expan))
-		return (-1);
-	arg->i++;
-	indx = 0;
-	while (str[arg->i] && (ft_isalnum(str[arg->i])
-			|| str[arg->i] == '_' || str[arg->i] == '?'))
-	{
-		if (indx < (int) sizeof(name) - 1)
-			name[indx++] = str[arg->i++];
-	}
-	name[indx] = '\0';
-	init_vdata(&data, expan, temp, name);
-	if (handle_value(mini, &data))
-		return (-1);
-	arg->name = ft_strdup(data.name);
-	arg->start += ft_strlen(arg->name) + 1;
-	free(arg->name);
-	return (arg->start);
-}
-
-int	tildes_home(t_shell *mini, char *str, char **expan, t_expand *arg)
-{
-	char	*temp;
-	char	*temp2;
-	char	*value;
-
-	temp2 = ft_strndup(&str[arg->start], arg->i - arg->start);
-	temp = ft_strjoin(*expan, temp2);
-	free(temp2);
-	free(*expan);
-	*expan = temp;
-	arg->i++;
-	value = get_value(mini->env, "HOME");
-	if (value)
-	{
-		temp = ft_strjoin(*expan, value);
-		free(*expan);
-		*expan = temp;
-		free(value);
-	}
-	return (arg->i);
-}
-
-static int	handle_new_expand(char *temp, char **expan)
-{
-	char	*new_expanded;
-
-	new_expanded = ft_strjoin(*expan, temp);
-	if (!new_expanded)
-	{
-		ft_putendl_fd("malloc failed", 2);
-		return (-1);
-	}
-	free(temp);
-	free(*expan);
-	*expan = new_expanded;
-	return (0);
-}
-
-static void	init_vdata(t_vdata *data, char **expan, char *temp, char *name)
-{
-	data->value = NULL;
-	data->expan = expan;
-	data->temp = temp;
-	data->name = name;
 }
