+++ sources/parser/split_input.c
@@ -6,7 +6,7 @@
 /*   By: nzharkev <nzharkev@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/07 10:05:25 by henbuska          #+#    #+#             */
-/*   Updated: 2024/12/21 19:48:03 by nzharkev         ###   ########.fr       */
+/*   Updated: 2024/12/23 18:43:28 by nzharkev         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,25 +15,46 @@
 int		split_input_by_pipes(char *input, t_shell *mini);
 char	*trim_whitespace(char *seg);
 
-static int	segment_handler(t_shell *mini, char *input, int start, int end, int *index)
+// static int	segment_handler(t_shell *mini, char *input, int s, int e, int *idx)
+// {
+// 	char	*trimmed;
+
+// 	trimmed = ft_strndup(input + s, e - s);
+// 	if (!trimmed)
+// 	{
+// 		ft_putendl_fd("Failed to allocate memory for trimmed string", 2);
+// 		return (1);
+// 	}
+// 	trimmed = trim_whitespace(trimmed);
+// 	mini->cmds[*idx]->seg = trimmed;
+// 	if (!mini->cmds[*idx])
+// 	{
+// 		ft_putendl_fd("No seg in struct", 2);
+// 		free(trimmed);
+// 		return (1);
+// 	}
+// 	(*idx)++;
+// 	return (0);
+// }
+
+static int	segment_handler(t_cmd *cmd, char *input, int s, int e)
 {
 	char	*trimmed;
 
-	trimmed = ft_strndup(input + start, end - start);
+	trimmed = ft_strndup(input + s, e - s);
 	if (!trimmed)
 	{
 		ft_putendl_fd("Failed to allocate memory for trimmed string", 2);
 		return (1);
 	}
 	trimmed = trim_whitespace(trimmed);
-	mini->cmds[*index]->seg = trimmed;
-	if (!mini->cmds[*index])
+	cmd->seg = trimmed;
+	if (!cmd->seg)
 	{
 		ft_putendl_fd("No seg in struct", 2);
 		free(trimmed);
 		return (1);
 	}
-	(*index)++;
 	return (0);
 }
 
@@ -50,14 +71,16 @@ int	split_input_by_pipes(char *input, t_shell *mini)
 	{
 		if (input[i] == '|' && !check_quotes(input, i))
 		{
-			if (segment_handler(mini, input, start, i, &index))
+			if (segment_handler(mini->cmds[index], input, start, i))
 				return (1);
 			start = i + 1;
+			index++;
 		}
 		i++;
 	}
-	if (segment_handler(mini, input,  start, i, &index))
+	if (segment_handler(mini->cmds[index], input,  start, i))
 		return (1);
+	index++;
 	return (0);
 }
